### Motivation
- The Windows MSVC build treats implicit narrowing conversions as errors
(C4244) and failed when assigning a `uint64_t` expression to the
`timestamp` field which is `uint32_t`, so the tests must use an explicit
cast to avoid the error.

### Description
- Add an explicit cast `static_cast<uint32_t>(...)` to the computed
timestamp assignments in `tests/unit/chainstate/test_chain.cpp` and
`tests/unit/chainstate/test_chainstate.cpp` to prevent implicit
`uint64_t` -> `uint32_t` narrowing.

### Testing
- Verified the changes by searching for the updated assignments with
`rg` and committed the changes; `rg` shows both occurrences now use
`static_cast<uint32_t>(...)` and the commit completed successfully.

------
[Codex
Task](https://chatgpt.com/codex/tasks/task_e_698a67f7d81483249feba5d7a8bc9c7d)

+++ tests/unit/chainstate/test_chain.cpp
@@ -48,7 +48,7 @@ Block CreateAndMineBlock(uint64_t height, const std::array<uint8_t, 32>& prev_ha
 
     block.header.version = 1;
     block.header.prev_block_hash = prev_hash;
-    block.header.timestamp = 1234567890 + height * 600;
+    block.header.timestamp = static_cast<uint32_t>(1234567890 + height * 600);
     block.header.bits = Difficulty::GetInitialBits();
     block.header.nonce = 0;
